refactor confirm form
