+++ .env
@@ -1,6 +1,6 @@
 REACT_APP_CHAIN_ID="3"
 REACT_APP_CHAIN_NAME_KEY="ROPSTEN"
-REACT_APP_NETWORK_URL="https://bsc-dataseed.binance.org/"
+REACT_APP_NETWORK_URL="https://ropsten.infura.io/v3/4bf032f2d38a4ed6bb975b80d6340847"
 
 SKIP_PREFLIGHT_CHECK=true
 STORYBOOK_MODE=true
