+++ .env.production
@@ -1,4 +1,5 @@
 REACT_APP_CHAIN_ID="56"
+REACT_APP_CHAIN_NAME_KEY="BSC"
 REACT_APP_NETWORK_URL="https://bsc-dataseed.binance.org/"
 REACT_APP_PORTIS_ID="c0e2bf01-4b08-4fd5-ac7b-8e26b58cd236"
 REACT_APP_FORTMATIC_KEY="pk_live_F937DF033A1666BF"
